+++ src/drivers/video/bochs/bochs.c
@@ -81,11 +81,9 @@ static int bochs_init(struct pci_slot_dev *pci_dev) {
 			* VBE_DISPI_MAX_BPP / 8, PAGE_SIZE());
 	struct fb_info *info;
 
-	if (MAP_FAILED == mmap_device_memory(mmap_base,
-				mmap_len,
-			       	PROT_READ|PROT_WRITE|PROT_NOCACHE,
-				MAP_FIXED,
-				(unsigned long) mmap_base)) {
+	if (MAP_FAILED == mmap_device_memory(mmap_base, mmap_len,
+	        PROT_READ | PROT_WRITE | PROT_NOCACHE, MAP_FIXED,
+	        (uintptr_t)mmap_base)) {
 		return -EIO;
 	}
